Rename substrate egress to actor resolution and stub real egress
The existing substrateEgress is just a frontend policy that resolves the
actor, so rename that to substrateEgressActorResolution (shorter names
welcome). I also stubbed out the new substrateEgress traffic policy that
will actually apply the policies from ateapi after fetching them.
