+++ crates/agentgateway/src/types/local.rs
@@ -2896,7 +2896,7 @@ struct LocalFrontendPolicies {
 	pub network_ext_authz: Option<crate::http::ext_authz::ExtAuthz>,
 	/// Validate the originating actor before accepting a CONNECT tunnel.
 	#[serde(default)]
-	pub substrate_egress: Option<crate::http::substrate::SubstrateEgress>,
+	pub substrate_egress: Option<crate::http::substrate::EgressActorResolution>,
 	/// Enable downstream PROXY protocol handling on this gateway or port, including
 	/// version matching and whether PROXY headers are required or optional.
 	#[serde(default, rename = "proxyProtocol", alias = "proxy")]
@@ -3000,6 +3000,9 @@ pub struct FilterOrPolicy {
 	/// Resolve Substrate actor hostnames for dynamic route backends on ingress.
 	#[serde(default)]
 	substrate_ingress: Option<crate::http::substrate::SubstrateIngress>,
+	/// Enforce the CONNECT-pinned effective Substrate egress policy.
+	#[serde(default)]
+	substrate_egress: Option<crate::http::substrate::SubstrateEgress>,
 	/// Modify request and response headers, bodies, or metadata.
 	#[serde(default)]
 	#[cfg_attr(
@@ -5194,7 +5197,10 @@ async fn split_frontend_policies(
 		);
 	}
 	if let Some(p) = substrate_egress {
-		add(FrontendPolicy::SubstrateEgress(p), "substrateEgress");
+		add(
+			FrontendPolicy::SubstrateEgressActorResolution(p),
+			"substrateEgress",
+		);
 	}
 	if let Some(p) = proxy_protocol {
 		add(FrontendPolicy::Proxy(p), "proxy");
@@ -5280,6 +5286,7 @@ pub(crate) async fn split_policies_for_target(
 		ext_authz,
 		ext_proc,
 		substrate_ingress,
+		substrate_egress,
 		buffer,
 		timeout,
 		retry,
@@ -5460,6 +5467,9 @@ pub(crate) async fn split_policies_for_target(
 	}
 	if let Some(p) = substrate_ingress {
 		route_policies.push(TrafficPolicy::SubstrateIngress(RequestPolicy::single(p)))
+	}
+	if let Some(p) = substrate_egress {
+		route_policies.push(TrafficPolicy::SubstrateEgress(RequestPolicy::single(p)))
 	}
 	if let Some(p) = local_rate_limit
 		&& !p.is_empty()
