+++ schema/config.json
@@ -2042,6 +2042,18 @@
           ],
           "default": null
         },
+        "substrateEgress": {
+          "description": "Enforce the CONNECT-pinned effective Substrate egress policy.",
+          "anyOf": [
+            {
+              "$ref": "#/$defs/SubstrateEgress"
+            },
+            {
+              "type": "null"
+            }
+          ],
+          "default": null
+        },
         "transformations": {
           "description": "Modify request and response headers, bodies, or metadata.",
           "anyOf": [
@@ -7693,6 +7705,81 @@
       },
       "additionalProperties": false
     },
+    "SubstrateEgress": {
+      "description": "Retrieves and enforces the current Substrate egress policy for each request.",
+      "type": "object",
+      "properties": {
+        "policies": {
+          "description": "Backend policies used when connecting to the service.",
+          "anyOf": [
+            {
+              "$ref": "#/$defs/SimpleLocalBackendPolicies"
+            },
+            {
+              "type": "null"
+            }
+          ]
+        }
+      },
+      "unevaluatedProperties": false,
+      "oneOf": [
+        {
+          "description": "Service reference. Service must be defined in the top level services list.",
+          "type": "object",
+          "properties": {
+            "service": {
+              "type": "object",
+              "properties": {
+                "name": {
+                  "description": "Name of the target Service, as defined in the top-level `services` list.",
+                  "$ref": "#/$defs/NamespacedHostname"
+                },
+                "port": {
+                  "description": "Port on the target Service to route to.",
+                  "type": "integer",
+                  "format": "uint16",
+                  "minimum": 0,
+                  "maximum": 65535
+                }
+              },
+              "additionalProperties": false,
+              "required": [
+                "name",
+                "port"
+              ]
+            }
+          },
+          "required": [
+            "service"
+          ]
+        },
+        {
+          "description": "Hostname or IP address",
+          "type": "object",
+          "properties": {
+            "host": {
+              "description": "Hostname or IP address",
+              "type": "string"
+            }
+          },
+          "required": [
+            "host"
+          ]
+        },
+        {
+          "type": "object",
+          "properties": {
+            "backend": {
+              "description": "Explicit backend reference. Backend must be defined in the top level backends list",
+              "type": "string"
+            }
+          },
+          "required": [
+            "backend"
+          ]
+        }
+      ]
+    },
     "LocalExplicitOrConditional_LocalTransformationConfig": {
       "anyOf": [
         {
@@ -9750,7 +9837,7 @@
           "description": "Validate the originating actor before accepting a CONNECT tunnel.",
           "anyOf": [
             {
-              "$ref": "#/$defs/SubstrateEgress"
+              "$ref": "#/$defs/EgressActorResolution"
             },
             {
               "type": "null"
@@ -10071,7 +10158,7 @@
       "description": "CEL authorization rules for downstream network connections.",
       "$ref": "#/$defs/RuleSet"
     },
-    "SubstrateEgress": {
+    "EgressActorResolution": {
       "description": "Validates an actor's identity before accepting a CONNECT tunnel.",
       "type": "object",
       "properties": {
